+++ pass-notification-service/pom.xml
@@ -195,6 +195,13 @@
             <scope>test</scope>
         </dependency>
 
+        <dependency>
+            <groupId>com.icegreen</groupId>
+            <artifactId>greenmail-junit5</artifactId>
+            <version>2.0.0</version>
+            <scope>test</scope>
+        </dependency>
+
     </dependencies>
 
    <build>
