+++ pass-notification-service/src/main/resources/application.properties
@@ -21,6 +21,8 @@
 #spring.activemq.broker-url=${activemq.broker.uri:tcp://${jms.host:localhost}:${jms.port:61616}}
 spring.jms.listener.concurrency=4
 spring.jms.cache.enabled=false
+
+# TODO make possible to pass in on command line as params, could just work
 spring.mail.host=localhost
 spring.mail.port=587
 spring.mail.username=
@@ -28,5 +30,4 @@ spring.mail.password=
 
 pass.notification.queue.event.name=event
 pass.notification.mode=DEMO
-pass.notification.mailer.debug=false
 pass.notification.configuration=classpath:/notification.json
