+++ pass-notification-service/src/test/resources/test-application.properties
@@ -21,8 +21,9 @@
 spring.activemq.broker-url=${activemq.broker.uri:tcp://${jms.host:localhost}:${jms.port:61616}}
 spring.jms.listener.concurrency=4
 spring.jms.listener.auto-startup=true
+spring.mail.protocol=smtp
 spring.mail.host=localhost
-spring.mail.port=587
+spring.mail.port=3025
 spring.mail.username=
 spring.mail.password=
 pass.notification.mode=DEMO
